refactor : Book만 따로 분리, service 레이어로 비즈니스로직 분리
