+++ book/src/main/java/untitled/infra/BookRepository.java
@@ -1,8 +1,9 @@
-package untitled.domain;
+package untitled.infra;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import untitled.domain.*;
+import untitled.domain.Book.Book;
 
 //<<< PoEAA / Repository
 @RepositoryRestResource(collectionResourceRel = "books", path = "books")
